[refactor] split the ApiActor to PublicApiActor and PrivateApiActor (and added ServerTime)
