+++ src/main/scala/io/ticofab/reactivekraken/model/package.scala
@@ -1,0 +1,5 @@
+package io.ticofab.reactivekraken
+
+package object model {
+  case class ServerTime(unixTime: Long, rfc113: String)
+}
