+++ src/main/scala/io/ticofab/reactivekraken/package.scala
@@ -1,0 +1,7 @@
+package io.ticofab
+
+package object reactivekraken {
+
+  trait Message
+  trait MessageResponse
+}
