+++ src/test/scala/io/ticofab/reactivekraken/KrakenPublicApiActorSpec.scala
@@ -19,8 +19,8 @@ package io.ticofab.reactivekraken
 import akka.actor.{ActorSystem, Props}
 import akka.http.scaladsl.model.HttpRequest
 import akka.testkit.{ImplicitSender, TestActorRef, TestKit, TestProbe}
+import io.ticofab.reactivekraken.KrakenPublicApiActor._
 import io.ticofab.reactivekraken.api.HttpRequestor
-import io.ticofab.reactivekraken.messages._
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
 import org.scalatest.mockito.MockitoSugar
@@ -30,7 +30,7 @@ import scala.concurrent.Future
 import scala.concurrent.duration._
 import scala.concurrent.ExecutionContext.Implicits.global
 
-class KrakenApiActorSpec extends TestKit(ActorSystem("KrakenApiActorSpec")) with ImplicitSender
+class KrakenPublicApiActorSpec extends TestKit(ActorSystem("KrakenApiActorSpec")) with ImplicitSender
   with WordSpecLike with Matchers with MockitoSugar {
 
   def nonceGenerator = () => System.currentTimeMillis
@@ -44,7 +44,7 @@ class KrakenApiActorSpec extends TestKit(ActorSystem("KrakenApiActorSpec")) with
         override def fireRequest(request: HttpRequest) = Future("mock")
       }
 
-      val apiActor: TestActorRef[KrakenApiActor] = TestActorRef(Props(spy(new KrakenApiActor(nonceGenerator) with MockHttpRequestor)))
+      val apiActor: TestActorRef[KrakenPublicApiActor] = TestActorRef(Props(spy(new KrakenPublicApiActor(nonceGenerator) with MockHttpRequestor)))
       val probe = TestProbe()
       probe.send(apiActor, GetCurrentAssets)
       probe.expectMsgType[CurrentAssets](3.seconds)
@@ -54,10 +54,28 @@ class KrakenApiActorSpec extends TestKit(ActorSystem("KrakenApiActorSpec")) with
     }
 
     "Not respond upon receiving a message it doesn't understand" in {
-      val testActor = TestActorRef[KrakenApiActor]
+      val testActor = TestActorRef[KrakenPublicApiActor]
       val probe = TestProbe()
       probe.send(testActor, "hello")
       probe.expectNoMsg(1.second)
+    }
+
+    "Can used from multiple context" in {
+      trait MockHttpRequestor extends HttpRequestor {
+        override def fireRequest(request: HttpRequest) = Future({Thread.sleep(1000); "mock"})
+      }
+
+      val apiActor: TestActorRef[KrakenPublicApiActor] = TestActorRef(Props(spy(new KrakenPublicApiActor(nonceGenerator) with MockHttpRequestor)))
+      val probe = TestProbe()
+      val probe2 = TestProbe()
+      probe.send(apiActor, GetCurrentAssets)
+      probe2.send(testActor, "hello")
+      probe.expectMsgType[CurrentAssets](3.seconds)
+
+      verify(apiActor.underlyingActor, times(1))
+        .fireRequest(ArgumentMatchers.any[HttpRequest])
+
+      probe2.expectNoMsg(5.second)
     }
 
   }
